+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     published
   * MiniReel Editor now creates MiniReels that will use the dynamic ad
     configuration
+  * The last splash theme/ratio the user selected is now stored in
+    LocalStorage/the user DB
   * Generated Wordpress Shortcode has been updated to the latest format
     based on feedback from USAToday